Port to Qt5
CMake cleanups coming

+++ src/core/regionselect.cpp
@@ -20,8 +20,8 @@
 
 #include "src/core/regionselect.h"
 
-#include <QtGui/QDesktopWidget>
-#include <QtGui/QApplication>
+#include <QDesktopWidget>
+#include <QApplication>
 
 RegionSelect::RegionSelect(Config *mainconf, QWidget *parent)
     :QWidget(parent)
@@ -161,7 +161,7 @@ void RegionSelect::drawBackGround()
     QRect txtRect = QApplication::desktop()->screenGeometry(QApplication::desktop()->primaryScreen());
     QString txtTip = QApplication::tr("Use your mouse to draw a rectangle to screenshot or exit pressing\nany key or using the right or middle mouse buttons.");
 
-    txtRect.setHeight(qRound(txtRect.height() / 10)); // rounded val of text rect height
+    txtRect.setHeight(qRound((float) (txtRect.height() / 10))); // rounded val of text rect height
 
     painter.setPen(QPen(Qt::red)); // ste message rect border color
     painter.setBrush(QBrush(QColor(255, 255, 255, 180), Qt::SolidPattern));
